+++ tests/next-app-dir-with-sentry-spotlight/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { HydrationOverlay } from "@builder.io/react-hydration-overlay";
+import { HydrationOverlay } from "@wayneintacart/react-hydration-overlay";
 
 const inter = Inter({ subsets: ["latin"] });