+++ tests/next-pages/next.config.js
@@ -1,6 +1,6 @@
 const {
   withHydrationOverlay,
-} = require("@builder.io/react-hydration-overlay/next");
+} = require("@wayneintacart/react-hydration-overlay/next");
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {