+++ tests/remix/vite.config.ts
@@ -1,7 +1,7 @@
 import { vitePlugin as remix } from "@remix-run/dev";
 import { defineConfig } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";
-import { withHydrationOverlayVite } from "@builder.io/react-hydration-overlay/vite";
+import { withHydrationOverlayVite } from "@wayneintacart/react-hydration-overlay/vite";
 
 declare module "@remix-run/node" {
   interface Future {